# Conflicts:
#	pom.xml
#	src/main/java/pl/wrapper/parking/facade/ParkingService.java
#	src/main/java/pl/wrapper/parking/facade/domain/ParkingController.java
#	src/main/java/pl/wrapper/parking/facade/domain/ParkingServiceImpl.java
#	src/main/java/pl/wrapper/parking/infrastructure/error/ErrorWrapper.java
#	src/main/java/pl/wrapper/parking/infrastructure/error/HandleResult.java
#	src/main/java/pl/wrapper/parking/infrastructure/exception/GlobalExceptionHandler.java
#	src/main/resources/application.properties
#	src/test/java/pl/wrapper/parking/facade/domain/ParkingControllerTests.java

+++ src/test/java/pl/wrapper/parking/infrastructure/error/ResultTest.java
@@ -27,9 +27,9 @@ class ResultTest {
 
     @Test
     void shouldReturnDummyBody() throws Exception {
-        Long id = 4L;
+        Integer id = 4;
 
-        Mockito.when(dummyService.dummyGetParkingBySymbol(id,true)).thenReturn(Result.success(id));
+        Mockito.when(dummyService.dummyGetParkingBySymbol(id, true)).thenReturn(Result.success(id));
 
         MvcResult mvcResult = mockMvc.perform(get("/id/{id}", id)) // add url and variables
                 .andReturn();
